2016-06-24 Self-Controller 修复添加头像会删除默认头像

+++ Application/Admin/Controller/SelfController.class.php
@@ -35,7 +35,9 @@ class SelfController extends BaseController {
 		//trace($info);
 		$adminUser = M('AdminUser');
 		$data = $adminUser -> where('id = %d',$id)->find();
-		unlink('Public/'.$data['img']);
+		if($data['img']!='/Images/User/default.png'){
+			unlink('Public/'.$data['img']);
+		}
 		$data['img'] = '/Images/User/'.$fname.'/'.$info['photo']['savename'];
 		$status = $adminUser -> save($data);
 		if ($info && $status) {// 上传成功
